Squashed commit of the following:
commit 95cd9d19bec03f7edce7b8afb2ab9a39ee4053d6
Author: spencer@primus <spencer.hallyburton@duke.edu>
Date:   Wed Aug 28 10:41:35 2024 -0400

    Fix fov estimator by adding centering

commit a770701516237b4bff8d894343679382bf2e75b8
Author: spencer@primus <spencer.hallyburton@duke.edu>
Date:   Tue Aug 27 20:19:49 2024 -0400

    Add robustness to A shape deficiency

commit bfcde8a4cd6ee0b8f9dc8f475703443ced98d95a
Author: spencer@primus <spencer.hallyburton@duke.edu>
Date:   Tue Aug 27 20:18:00 2024 -0400

    Update mmdet3d

commit 7d0d649b9e2b8b652e63392b12ec0306e26cb108
Author: spencer@primus <spencer.hallyburton@duke.edu>
Date:   Mon Aug 26 10:38:08 2024 -0400

    Add metrics and some simple tests

+++ tests/geometry/test_fov.py
@@ -35,8 +35,8 @@ from utilities import get_test_sensor_data
 
 def test_lidar_concave_hull():
     hull = pc.concave_hull_bev()
-    assert hull.check_point(np.array([0, 0]))
-    assert not hull.check_point(np.array([200, 0]))
+    assert hull.check_point(np.array([0, 0], dtype=float))
+    assert not hull.check_point(np.array([200, 0], dtype=float))
 
 
 def test_make_fov_from_pc_hull():
@@ -46,7 +46,7 @@ def test_make_fov_from_pc_hull():
 
 def test_point_in_fov():
     fov = pc.concave_hull_bev()
-    in_fov = points_in_fov(np.array([0, 0, 0]), fov)
+    in_fov = points_in_fov(np.array([0, 0, 0], dtype=float), fov)
     assert in_fov
